refactor(structure): separate HomePage from App; convert App to layout component
- Move homepage content to pages/HomePage.tsx
- App.tsx now serves as layout with header and routing logic
- Cleaner separation of concerns for routing and page content
